+++ xenon-common/src/test/java/com/vmware/xenon/services/common/TestQueryTaskUtils.java
@@ -189,6 +189,7 @@ public class TestQueryTaskUtils {
         expectedPropertyNames.add(ExampleServiceState.FIELD_NAME_ID);
         expectedPropertyNames.add(ExampleServiceState.FIELD_NAME_REQUIRED);
         expectedPropertyNames.add(ExampleServiceState.FIELD_NAME_KEY_VALUES);
+        expectedPropertyNames.add(ExampleServiceState.FIELD_NAME_IS_FROM_MIGRATION);
         expectedPropertyNames.add("tags.item");
 
         assertEquals(expectedPropertyNames, queryPropertyNames);